refactor(lsp): extract LSP layer into dedicated package
LSP-сервисы и холдеры лежали в корневом пакете languageserver вперемешку с
launcher'ом и прочим. Выделены в отдельный пакет lsp/ (симметрично mcp/):
сервисы BSLLanguageServer/BSLTextDocumentService/BSLWorkspaceService и холдеры
ClientCapabilitiesHolder/LanguageClientHolder (это чистые понятия LSP) — вместе
со своими тестами. Импорты обновлены по всему проекту; launcher (BSLLSPLauncher,
Main-Class) и не-LSP классы остаются в корне.

ArchUnit: добавлены слои Lsp и Mcp, правило для providers ужесточено до реальных
потребителей — providers вызывается только из голов lsp/cli/mcp (без layers
«про запас»).
