+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/aop/SentryAspect.java
@@ -21,7 +21,7 @@
  */
 package com.github._1c_syntax.bsl.languageserver.aop;
 
-import com.github._1c_syntax.bsl.languageserver.LanguageClientHolder;
+import com.github._1c_syntax.bsl.languageserver.lsp.LanguageClientHolder;
 import com.github._1c_syntax.bsl.languageserver.configuration.GlobalLanguageServerConfiguration;
 import com.github._1c_syntax.bsl.languageserver.configuration.LanguageServerConfiguration;
 import com.github._1c_syntax.bsl.languageserver.infrastructure.WorkspaceContextHolder;
