+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/lsp/LanguageClientHolder.java
@@ -19,7 +19,7 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with BSL Language Server.
  */
-package com.github._1c_syntax.bsl.languageserver;
+package com.github._1c_syntax.bsl.languageserver.lsp;
 
 import lombok.NoArgsConstructor;
 import org.eclipse.lsp4j.services.LanguageClient;
